Added portfolio link
Changed the about tab to portfolio and it now links to the portfolio site. Since the tab will just open a new page in the blog (and this has to happen due to the template) the redirecting is a script that runs in the portfolio page. Fortunately, browsers seem to be smart enough to not include the redirection page in the page history. So if the viewer presses the back button, it should take them to the page they were on before they clicked the about button.